# WIP: Improve FindMeshBlock performance