--- a/src/mesh/meshblock.cpp
+++ b/src/mesh/meshblock.cpp
@@ -63,8 +63,8 @@ MeshBlock::MeshBlock(const int n_side, const int ndim)
 
 MeshBlock::MeshBlock(int igid, int ilid, LogicalLocation iloc, RegionSize input_block,
                      BoundaryFlag *input_bcs, Mesh *pm, ParameterInput *pin,
-                     Properties_t &properties, Packages_t &packages, int igflag,
-                     bool ref_flag)
+                     Properties_t const &properties, Packages_t const &packages,
+                     int igflag, bool ref_flag)
     : exec_space(DevExecSpace()), pmy_mesh(pm), loc(iloc), block_size(input_block),
       gid(igid), lid(ilid), gflag(igflag), properties(properties), packages(packages),
       prev(nullptr), next(nullptr), new_block_dt_{}, new_block_dt_hyperbolic_{},
@@ -147,9 +147,9 @@ MeshBlock::MeshBlock(int igid, int ilid, LogicalLocation iloc, RegionSize input_
 // MeshBlock constructor for restarts
 #if 0
 MeshBlock::MeshBlock(int igid, int ilid, Mesh *pm, ParameterInput *pin,
-                     Properties_t &properties, Packages_t &packages, LogicalLocation iloc,
-                     RegionSize input_block, BoundaryFlag *input_bcs, double icost,
-                     char *mbdata, int igflag)
+                     Properties_t const &properties, Packages_t const &packages,
+                     LogicalLocation iloc, RegionSize input_block,
+                     BoundaryFlag *input_bcs, double icost, char *mbdata, int igflag)
     : pmy_mesh(pm), loc(iloc), block_size(input_block), gid(igid), lid(ilid),
       gflag(igflag), nuser_out_var(), properties(properties), packages(packages),
       prev(nullptr), next(nullptr), new_block_dt_{}, new_block_dt_hyperbolic_{},