--- a/tst/performance/CMakeLists.txt
+++ b/tst/performance/CMakeLists.txt
@@ -16,7 +16,10 @@
 ## the public, perform publicly and display publicly, and to permit others to do so.
 ##========================================================================================
 
-add_executable(performance_tests test_container_iterator.cpp)
+add_executable(
+    performance_tests
+    test_container_iterator.cpp
+    test_find_block.cpp)
 target_link_libraries(performance_tests PRIVATE parthenon catch2_define Kokkos::kokkos)
 lint_target(performance_tests)