## PR Summary

This PR is intended to improve `FindMeshBlock` performance. Initial reports are not good... These numbers are for finding the last block for the following square meshes

```
benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Find Last Block 16x16                          100          4266     1.7064 ms 
                                         4.5963 ns    4.58576 ns    4.62287 ns 
                                      0.0739888 ns 0.00114742 ns   0.134083 ns 
                                                                               
Find Last Block 64x64                          100            91     1.9747 ms 
                                         216.37 ns    215.884 ns    217.592 ns 
                                        3.54582 ns   0.930452 ns    6.31153 ns 
                                                                               
Find Last Block 128x128                        100            11     2.0955 ms 
                                        1.89586 us    1.89168 us    1.90653 us 
                                        30.0121 ns    3.42593 ns    54.7218 ns 
                                                                               
Find Last Block 256x256                        100             3     2.6199 ms 
                                        8.70636 us    8.68915 us    8.74003 us 
                                        118.343 ns     65.985 ns    184.963 ns 
                                                                               
Find Last Block 512x512                        100             1     5.2201 ms 
                                        52.3517 us    52.2113 us    52.7356 us 
                                        1.09039 us     501.03 ns    2.28704 us
```

## PR Checklist
- [x] Code passes cpplint
- [x] New features are documented.
- [x] Adds a test for any bugs fixed. Adds tests for new features.
- [x] Code is formatted